\name{Barley5H}
\alias{Barley5H}
\docType{data}
\title{
Linkage maps for barley chromosome 5H
}
\description{
SNP marker linkage maps for barley chromosome 5H from the Steptoe x Morex (SM), Oregon Wolfe Barley (OWB), 
Morex x Barke (MB), and Haruno Nijo x OUH602 (HO) populations.
}
\usage{data(Barley5H)}
\details{
Barley5H is a list with 4 data frames (SM, OWB, MB, HO), representing 4 linkage maps.
The first column of each data frame contains the marker names, and the second
column contains the linkage map positions.
}
\references{
Close et al. 2009. Development and implementation of high-throughput SNP genotyping in barley. 
BMC Genomics 10:582. \url{http://www.biomedcentral.com/1471-2164/10/582}
}

